Refactored code so that peaks are represented as floats and the cluster charge is an int as the most common charge All unit tests pass - some still marked as Known to fail 2) Almost all warnings eliminated with the levels in slewis' intellij

+++ trunk/src/main/java/uk/ac/ebi/pride/spectracluster/cluster/ISpectralCluster.java
@@ -24,7 +24,7 @@ public interface ISpectralCluster extends Equivalent<ISpectralCluster>,Comparabl
      *
      * @return
      */
-    public double getPrecursorMz();
+    public float getPrecursorMz();
 
 
     /**
@@ -32,7 +32,7 @@ public interface ISpectralCluster extends Equivalent<ISpectralCluster>,Comparabl
      *
      * @return
      */
-    public double getPrecursorCharge();
+    public int getPrecursorCharge();
 
 
     /**
